chore: refactor to reduce diff
