--- internal/functions/serve/service_test.go
@@ -1,455 +1,0 @@
-package serve
-
-import (
-	"context"
-	"encoding/json"
-	"net/http"
-	"testing"
-	"time"
-
-	"github.com/h2non/gock"
-	"github.com/spf13/afero"
-	"github.com/spf13/viper"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-	"github.com/supabase/cli/internal/testing/apitest"
-	"github.com/supabase/cli/internal/utils"
-	"github.com/supabase/cli/pkg/cast"
-)
-
-func TestParseEnvFile(t *testing.T) {
-	// Save original CurrentDirAbs
-	originalCurrentDirAbs := utils.CurrentDirAbs
-	defer func() {
-		utils.CurrentDirAbs = originalCurrentDirAbs
-	}()
-
-	t.Run("parses env file successfully", func(t *testing.T) {
-		setup := NewTestSetup(t)
-		defer setup.Cleanup()
-
-		envContent := `DATABASE_URL=postgresql://localhost:5432/test
-API_KEY=secret123
-DEBUG=true`
-		envPath := "/project/.env"
-		setup.SetupEnvFile(envPath, envContent)
-
-		env, err := parseEnvFile(envPath, setup.Fsys)
-		assert.NoError(t, err)
-		assert.Contains(t, env, "DATABASE_URL=postgresql://localhost:5432/test")
-		assert.Contains(t, env, "API_KEY=secret123")
-		assert.Contains(t, env, "DEBUG=true")
-	})
-
-	t.Run("uses fallback env file when path is empty", func(t *testing.T) {
-		setup := NewTestSetup(t)
-		defer setup.Cleanup()
-
-		envContent := `FALLBACK_VAR=fallback_value`
-		setup.SetupEnvFile("", envContent)
-
-		env, err := parseEnvFile("", setup.Fsys)
-		assert.NoError(t, err)
-		assert.Contains(t, env, "FALLBACK_VAR=fallback_value")
-	})
-
-	t.Run("returns error when file doesn't exist", func(t *testing.T) {
-		setup := NewTestSetup(t)
-		defer setup.Cleanup()
-
-		envPath := "/project/nonexistent.env"
-		env, err := parseEnvFile(envPath, setup.Fsys)
-		assert.Error(t, err)
-		assert.Nil(t, env)
-		assert.Contains(t, err.Error(), "failed to open env file")
-	})
-}
-
-func TestPopulatePerFunctionConfigs(t *testing.T) {
-	// Save original values
-	originalFunctionsDir := utils.FunctionsDir
-	defer func() {
-		utils.FunctionsDir = originalFunctionsDir
-	}()
-
-	t.Run("populates function configs successfully", func(t *testing.T) {
-		setup := NewTestSetup(t)
-		defer setup.Cleanup()
-
-		utils.FunctionsDir = "functions"
-		setup.SetupFunction("hello", "export default () => 'hello'")
-		setup.SetupConfigWithFunctions()
-
-		binds, configString, err := populatePerFunctionConfigs("/project", "", cast.Ptr(false), setup.Fsys)
-		assert.NoError(t, err)
-		assert.NotEmpty(t, binds)
-		assert.NotEmpty(t, configString)
-
-		var config map[string]interface{}
-		err = json.Unmarshal([]byte(configString), &config)
-		assert.NoError(t, err)
-		assert.Contains(t, config, "hello")
-	})
-
-	t.Run("handles function config creation", func(t *testing.T) {
-		setup := NewTestSetup(t)
-		defer setup.Cleanup()
-
-		utils.FunctionsDir = "functions"
-		setup.SetupFunction("enabled", "export default () => 'enabled'")
-
-		_, configString, err := populatePerFunctionConfigs("/project", "", nil, setup.Fsys)
-		assert.NoError(t, err)
-
-		var resultConfig map[string]interface{}
-		err = json.Unmarshal([]byte(configString), &resultConfig)
-		assert.NoError(t, err)
-		assert.Contains(t, resultConfig, "enabled")
-
-		enabledConfig := resultConfig["enabled"].(map[string]interface{})
-		assert.Contains(t, enabledConfig, "entrypointPath")
-		assert.Contains(t, enabledConfig, "verifyJWT")
-	})
-
-	t.Run("handles import map path", func(t *testing.T) {
-		setup := NewTestSetup(t)
-		defer setup.Cleanup()
-
-		utils.FunctionsDir = "functions"
-		setup.SetupFunction("hello", "export default () => 'hello'")
-		setup.SetupImportMap("import_map.json", "{}")
-
-		binds, configString, err := populatePerFunctionConfigs("/project", "import_map.json", nil, setup.Fsys)
-		assert.NoError(t, err)
-		assert.NotEmpty(t, binds)
-		assert.NotEmpty(t, configString)
-	})
-
-	t.Run("returns empty config when no functions exist", func(t *testing.T) {
-		setup := NewTestSetup(t)
-		defer setup.Cleanup()
-
-		utils.FunctionsDir = "functions"
-		require.NoError(t, setup.Fsys.MkdirAll("functions", 0755))
-
-		_, configString, err := populatePerFunctionConfigs("/project", "", nil, setup.Fsys)
-		assert.NoError(t, err)
-
-		var resultConfig map[string]interface{}
-		err = json.Unmarshal([]byte(configString), &resultConfig)
-		assert.NoError(t, err)
-		assert.Empty(t, resultConfig)
-	})
-}
-
-func TestServeFunctions(t *testing.T) {
-	// Save original values
-	originalConfig := utils.Config
-	originalDebug := viper.Get("DEBUG")
-	originalFunctionsDir := utils.FunctionsDir
-	defer func() {
-		utils.Config = originalConfig
-		viper.Set("DEBUG", originalDebug)
-		utils.FunctionsDir = originalFunctionsDir
-	}()
-
-	t.Run("returns error on env file parsing failure", func(t *testing.T) {
-		// Setup in-memory fs
-		fsys := afero.NewMemMapFs()
-		// Test with nonexistent file to trigger error
-
-		// Test function
-		err := ServeFunctions(context.Background(), "nonexistent.env", nil, "", "postgresql://localhost:5432/test", RuntimeOption{}, fsys)
-		assert.Error(t, err)
-		assert.Contains(t, err.Error(), "failed to open env file")
-	})
-
-	t.Run("returns error on function config failure", func(t *testing.T) {
-		// Setup config
-		utils.Config.Auth.AnonKey.Value = "anon_key"
-		utils.Config.Auth.ServiceRoleKey.Value = "service_role_key"
-		utils.Config.Auth.JwtSecret.Value = "jwt_secret"
-		utils.Config.Api.Port = 8000
-		utils.Config.EdgeRuntime.Policy = "permissive"
-		utils.KongAliases = []string{"supabase_kong_test"}
-
-		// Setup in-memory fs with invalid functions directory
-		fsys := afero.NewMemMapFs()
-		utils.FunctionsDir = "nonexistent"
-
-		// Test function
-		err := ServeFunctions(context.Background(), "", nil, "", "postgresql://localhost:5432/test", RuntimeOption{}, fsys)
-		assert.Error(t, err)
-	})
-}
-
-func TestManageFunctionServices(t *testing.T) {
-	// Save original values
-	originalConfig := utils.Config
-	defer func() {
-		utils.Config = originalConfig
-	}()
-
-	setupMockConfig := func() {
-		utils.Config.Auth.AnonKey.Value = "anon_key"
-		utils.Config.Auth.ServiceRoleKey.Value = "service_role_key"
-		utils.Config.Auth.JwtSecret.Value = "jwt_secret"
-		utils.Config.Api.Port = 8000
-		utils.Config.EdgeRuntime.Policy = "permissive"
-		utils.Config.EdgeRuntime.Image = "supabase/edge-runtime:test"
-		utils.KongAliases = []string{"supabase_kong_test"}
-		utils.EdgeRuntimeId = "test-edge-runtime"
-	}
-
-	t.Run("manages function services successfully", func(t *testing.T) {
-		setupMockConfig()
-
-		// Setup in-memory fs
-		fsys := afero.NewMemMapFs()
-
-		// Setup mock docker
-		require.NoError(t, apitest.MockDocker(utils.Docker))
-		defer gock.OffAll()
-		// Mock container removal
-		gock.New(utils.Docker.DaemonHost()).
-			Delete("/v" + utils.Docker.ClientVersion() + "/containers/" + utils.EdgeRuntimeId).
-			Reply(http.StatusOK)
-		// Mock container start
-		apitest.MockDockerStart(utils.Docker, utils.GetRegistryImageUrl(utils.Config.EdgeRuntime.Image), utils.EdgeRuntimeId)
-		// Mock log streaming
-		require.NoError(t, apitest.MockDockerLogs(utils.Docker, utils.EdgeRuntimeId, "service started"))
-
-		// Create error channel
-		errChan := make(chan error, 1)
-
-		// Test function
-		cancel, logsDone, err := manageFunctionServices(
-			context.Background(),
-			"",
-			nil,
-			"",
-			"postgresql://localhost:5432/test",
-			RuntimeOption{},
-			fsys,
-			errChan,
-		)
-
-		assert.NoError(t, err)
-		assert.NotNil(t, cancel)
-		assert.NotNil(t, logsDone)
-
-		// Cancel and wait for cleanup if we got valid results
-		if cancel != nil {
-			cancel()
-			select {
-			case <-logsDone:
-				// Expected
-			case <-time.After(1 * time.Second):
-				t.Error("Expected logsDone to close after cancel")
-			}
-		}
-
-		assert.Empty(t, apitest.ListUnmatchedRequests())
-	})
-
-	t.Run("returns error when ServeFunctions fails", func(t *testing.T) {
-		setupMockConfig()
-
-		// Setup in-memory fs with nonexistent env file to trigger error
-		fsys := afero.NewMemMapFs()
-
-		// Setup mock docker for container removal
-		require.NoError(t, apitest.MockDocker(utils.Docker))
-		defer gock.OffAll()
-		gock.New(utils.Docker.DaemonHost()).
-			Delete("/v" + utils.Docker.ClientVersion() + "/containers/" + utils.EdgeRuntimeId).
-			Reply(http.StatusOK)
-
-		// Create error channel
-		errChan := make(chan error, 1)
-
-		// Test function
-		cancel, logsDone, err := manageFunctionServices(
-			context.Background(),
-			"nonexistent.env",
-			nil,
-			"",
-			"postgresql://localhost:5432/test",
-			RuntimeOption{},
-			fsys,
-			errChan,
-		)
-
-		assert.Error(t, err)
-		assert.Nil(t, cancel)
-		assert.Nil(t, logsDone)
-		assert.Contains(t, err.Error(), "Failed to serve functions")
-	})
-
-	t.Run("handles context cancellation", func(t *testing.T) {
-		setupMockConfig()
-
-		// Setup in-memory fs
-		fsys := afero.NewMemMapFs()
-
-		// Setup mock docker
-		require.NoError(t, apitest.MockDocker(utils.Docker))
-		defer gock.OffAll()
-		// Mock container removal
-		gock.New(utils.Docker.DaemonHost()).
-			Delete("/v" + utils.Docker.ClientVersion() + "/containers/" + utils.EdgeRuntimeId).
-			Reply(http.StatusOK)
-		// Mock container start
-		apitest.MockDockerStart(utils.Docker, utils.GetRegistryImageUrl(utils.Config.EdgeRuntime.Image), utils.EdgeRuntimeId)
-		// Mock log streaming that will be cancelled
-		gock.New(utils.Docker.DaemonHost()).
-			Get("/v"+utils.Docker.ClientVersion()+"/containers/"+utils.EdgeRuntimeId+"/logs").
-			Reply(http.StatusOK).
-			SetHeader("Content-Type", "application/vnd.docker.raw-stream").
-			Delay(100 * time.Millisecond) // Delay to allow cancellation
-
-		// Create context that will be cancelled
-		ctx, cancel := context.WithCancel(context.Background())
-		errChan := make(chan error, 1)
-
-		// Test function
-		serviceCancel, logsDone, err := manageFunctionServices(
-			ctx,
-			"",
-			nil,
-			"",
-			"postgresql://localhost:5432/test",
-			RuntimeOption{},
-			fsys,
-			errChan,
-		)
-
-		assert.NoError(t, err)
-		assert.NotNil(t, serviceCancel)
-		assert.NotNil(t, logsDone)
-
-		// Cancel the context
-		cancel()
-
-		// Wait for logs to be done
-		select {
-		case <-logsDone:
-			// Expected
-		case <-time.After(1 * time.Second):
-			t.Error("Expected logsDone to close after context cancellation")
-		}
-
-		// No error should be sent to errChan for context cancellation
-		select {
-		case err := <-errChan:
-			t.Errorf("Unexpected error in errChan: %v", err)
-		default:
-			// Expected - no error for context cancellation
-		}
-	})
-
-	t.Run("sends error to channel on docker log streaming failure", func(t *testing.T) {
-		setupMockConfig()
-
-		// Setup in-memory fs
-		fsys := afero.NewMemMapFs()
-
-		// Setup mock docker
-		require.NoError(t, apitest.MockDocker(utils.Docker))
-		defer gock.OffAll()
-		// Mock container removal
-		gock.New(utils.Docker.DaemonHost()).
-			Delete("/v" + utils.Docker.ClientVersion() + "/containers/" + utils.EdgeRuntimeId).
-			Reply(http.StatusOK)
-		// Mock container start
-		apitest.MockDockerStart(utils.Docker, utils.GetRegistryImageUrl(utils.Config.EdgeRuntime.Image), utils.EdgeRuntimeId)
-		// Mock log streaming failure
-		gock.New(utils.Docker.DaemonHost()).
-			Get("/v" + utils.Docker.ClientVersion() + "/containers/" + utils.EdgeRuntimeId + "/logs").
-			Reply(http.StatusServiceUnavailable)
-
-		// Create error channel
-		errChan := make(chan error, 1)
-
-		// Test function
-		serviceCancel, logsDone, err := manageFunctionServices(
-			context.Background(),
-			"",
-			nil,
-			"",
-			"postgresql://localhost:5432/test",
-			RuntimeOption{},
-			fsys,
-			errChan,
-		)
-
-		assert.NoError(t, err)
-		assert.NotNil(t, serviceCancel)
-		assert.NotNil(t, logsDone)
-
-		// Wait for error or timeout
-		select {
-		case err := <-errChan:
-			assert.Error(t, err)
-			assert.Contains(t, err.Error(), "Docker log streaming error")
-		case <-time.After(1 * time.Second):
-			t.Error("Expected error in errChan")
-		}
-
-		// Wait for cleanup
-		select {
-		case <-logsDone:
-			// Expected
-		case <-time.After(1 * time.Second):
-			t.Error("Expected logsDone to close")
-		}
-	})
-
-	t.Run("handles full error channel gracefully", func(t *testing.T) {
-		setupMockConfig()
-
-		// Setup in-memory fs
-		fsys := afero.NewMemMapFs()
-
-		// Setup mock docker
-		require.NoError(t, apitest.MockDocker(utils.Docker))
-		defer gock.OffAll()
-		// Mock container removal
-		gock.New(utils.Docker.DaemonHost()).
-			Delete("/v" + utils.Docker.ClientVersion() + "/containers/" + utils.EdgeRuntimeId).
-			Reply(http.StatusOK)
-		// Mock container start
-		apitest.MockDockerStart(utils.Docker, utils.GetRegistryImageUrl(utils.Config.EdgeRuntime.Image), utils.EdgeRuntimeId)
-		// Mock log streaming failure
-		gock.New(utils.Docker.DaemonHost()).
-			Get("/v" + utils.Docker.ClientVersion() + "/containers/" + utils.EdgeRuntimeId + "/logs").
-			Reply(http.StatusServiceUnavailable)
-
-		// Create full error channel (capacity 0)
-		errChan := make(chan error)
-
-		// Test function
-		serviceCancel, logsDone, err := manageFunctionServices(
-			context.Background(),
-			"",
-			nil,
-			"",
-			"postgresql://localhost:5432/test",
-			RuntimeOption{},
-			fsys,
-			errChan,
-		)
-
-		assert.NoError(t, err)
-		assert.NotNil(t, serviceCancel)
-		assert.NotNil(t, logsDone)
-
-		// Wait for cleanup - should not hang even with full error channel
-		select {
-		case <-logsDone:
-			// Expected
-		case <-time.After(1 * time.Second):
-			t.Error("Expected logsDone to close even with full error channel")
-		}
-	})
-}
